# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://rrn24.techchantier.com/tutor-finder/public/api'
paths:
...
  /api/connection-requests:
    get:
      summary: 'Get all connection requests for the authenticated user'
      operationId: getAllConnectionRequestsForTheAuthenticatedUser
      description: "Returns connection requests either sent or received by the user,\ndepending on whether they are a learner or tutor."
      parameters:
        -
          in: query
          name: status
          description: 'Filter by request status (pending/accepted/rejected)'
          example: reiciendis
          required: false
          schema:
            type: string
            description: 'Filter by request status (pending/accepted/rejected)'
            example: reiciendis
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      learner:
                        id: 2
                        name: 'John Doe'
                      tutor:
                        id: 3
                        name: 'Jane Smith'
                      message: "I'd like to connect for mathematics tutoring"
                      status: pending
                      period_type: weekly
                      amount_paid: 5000
                      payment_completed: true
                      created_at: '2023-06-15T14:30:00.000Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        learner:
                          id: 2
                          name: 'John Doe'
                        tutor:
                          id: 3
                          name: 'Jane Smith'
                        message: "I'd like to connect for mathematics tutoring"
                        status: pending
                        period_type: weekly
                        amount_paid: 5000
                        payment_completed: true
                        created_at: '2023-06-15T14:30:00.000Z'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        learner:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 2
                            name:
                              type: string
                              example: 'John Doe'
                        tutor:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 3
                            name:
                              type: string
                              example: 'Jane Smith'
                        message:
                          type: string
                          example: "I'd like to connect for mathematics tutoring"
                        status:
                          type: string
                          example: pending
                        period_type:
                          type: string
                          example: weekly
                        amount_paid:
                          type: integer
                          example: 5000
                        payment_completed:
                          type: boolean
                          example: true
                        created_at:
                          type: string
                          example: '2023-06-15T14:30:00.000Z'
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: accepted
                  nullable: false
                  enum:
                    - pending
                    - accepted
                    - rejected
      security: []
    post:
      summary: 'Create Connection Request'
      operationId: createConnectionRequest
      description: "Creates a new connection request from a learner to a tutor.\nOnly authenticated learners can create connection requests."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    learner:
                      id: 2
                      name: 'John Doe'
                    tutor:
                      id: 3
                      name: 'Jane Smith'
                    message: 'I would like help with advanced calculus topics'
                    status: pending
                    period_type: weekly
                    amount_paid: 5000
                    payment_completed: false
                    created_at: '2023-06-15T14:30:00.000Z'
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      learner:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 2
                          name:
                            type: string
                            example: 'John Doe'
                      tutor:
                        type: object
                        properties:
                          id:
                            type: integer
                            example: 3
                          name:
                            type: string
                            example: 'Jane Smith'
                      message:
                        type: string
                        example: 'I would like help with advanced calculus topics'
                      status:
                        type: string
                        example: pending
                      period_type:
                        type: string
                        example: weekly
                      amount_paid:
                        type: integer
                        example: 5000
                      payment_completed:
                        type: boolean
                        example: false
                      created_at:
                        type: string
                        example: '2023-06-15T14:30:00.000Z'
        400:
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: ''
                    type: object
                    example:
                      message: 'You already have a pending request for this tutor.'
                    properties:
                      message:
                        type: string
                        example: 'You already have a pending request for this tutor.'
                  -
                    description: ''
                    type: object
                    example:
                      message: 'Selected user is not a tutor'
                    properties:
                      message:
                        type: string
                        example: 'Selected user is not a tutor'
                  -
                    description: ''
                    type: object
                    example:
                      message: 'Tutor has not set price for this period'
                    properties:
                      message:
                        type: string
                        example: 'Tutor has not set price for this period'
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutor_id:
                  type: integer
                  description: 'The ID of the tutor to connect with.'
                  example: 3
                  nullable: false
                message:
                  type: string
                  description: 'A message to the tutor explaining the connection request.'
                  example: 'I would like help with advanced calculus topics.'
                  nullable: false
                learner_whatsapp:
                  type: string
                  description: "The learner's WhatsApp number for contact purposes."
                  example: '+1234567890'
                  nullable: false
                period_type:
                  type: string
                  description: 'The billing period for the connection (weekly or monthly).'
                  example: weekly
                  nullable: false
              required:
                - tutor_id
                - message
                - learner_whatsapp
                - period_type
      security: []
...
  '/api/connection-requests/{id}/accept':
    post:
      summary: 'Accept Connection Request'
      operationId: acceptConnectionRequest
      description: "Allows a tutor to accept a connection request that has been paid for.\nSets the connection start and end dates and updates status to accepted."
      parameters: []
      responses:
        400:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"data\": {\n    \"id\": 1,\n    \"learner\": {\n      \"id\": 2,\n      \"name\": \"John Doe\"\n    },\n    \"tutor\": {\n      \"id\": 3,\n      \"name\": \"Jane Smith\"\n    },\n    \"message\": \"I would like help with advanced calculus topics\",\n    \"status\": \"accepted\",\n    \"period_type\": \"weekly\",\n    \"amount_paid\": 5000,\n    \"payment_completed\": true,\n    \"tutor_whatsapp\": \"+1234567890\",\n    \"start_date\": \"2023-06-15T14:30:00.000Z\",\n    \"end_date\": \"2023-06-22T14:30:00.000Z\",\n    \"created_at\": \"2023-06-15T14:30:00.000Z\"\n  }\n}\n\n* @response {\n  \"message\": \"Cannot accept request. Payment not completed.\"\n}"
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tutor_whatsapp:
                  type: string
                  description: "The tutor's WhatsApp number for contact purposes."
                  example: '+1234567890'
                  nullable: false
              required:
                - tutor_whatsapp
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection request.'
        example: 1
        required: true
        schema:
          type: integer
...